[005][fix] Changing to long type for big numbers

+++ SumOrProduct/Product.cs
@@ -3,8 +3,8 @@ using System;
 namespace SumOrProduct{
     class Product{
 
-        public static void prod(int num){
-            int ans = 1;
+        public static void prod(long num){
+            long ans = 1;
             for (int i = 1; i < num + 1; i++){
                 ans = ans * i;
             }
